Example moving logic to services.
First step is to identify what logic can be moved to which services. This case we can simply move the logic to marketplace manager. As the original packet had to return an success status (true / false) we make this the result of our method. We pass the Habbo who is doing this action so that the inventory can also be updated.