Step 2: Update queries to dapper.
This might be a bit complicated and you might have to change some stuff around. In this case we needed some extra fields so we added them for now to MarketOffer and queried that object. A method GetOffer has been extracted to query for an offer. Another method DeleteOffer has been extracted to delete the offer from the database.